fix: resolve compilation errors in autosteer and stack monitoring
Compilation Fixes:
- autosteer.cpp:38: Add static_cast to fix type mismatch between
  abs() return type (int) and MAX_PWM_VALUE (uint8_t)
- main.cpp:56,62,68,74,80: Change warning() to warningf() to support
  printf-style formatting in stack monitoring

These errors were introduced after linter modifications to the codebase.

+++ src/autosteer/autosteer.cpp
@@ -35,7 +35,7 @@ void handler() {
     auto control_out = calcSteeringPID(steerAngleError); //do the pid
 
     // Properly limit the PWM value to valid range
-    uint8_t pwm   = min(abs(control_out), MAX_PWM_VALUE);
+    uint8_t pwm   = min(static_cast<uint8_t>(abs(control_out)), MAX_PWM_VALUE);
     bool reversed = control_out < 0;
 
     if (steerEnable) {
